RD-781 Reference docs and examples (#185)
* Upgrade to MapTiler client v2.3.0 (#180)

* feat(package.json): upgrade maptiler-client-js to 2.3.0

* 3.1.0

* chore: update readme and changelog

---------


* RD-781 Add DOCTYPE to examples / demos to avoid quirks mode

* RD-781 Export AnimatedRouteLayer from index

* RD-781 Create Demo boilerplate

* RD-781 Update readme and docs.

* RD-781 Fix typo in readme

* RD-781 Fix additional typo in readme

* RD-781 Add additional info regarding exclusioonary usage of AnimatedRouteLayer

---------
