+++ demos/02-sky-day.html
@@ -1,5 +1,5 @@
+<!DOCTYPE html>
 <html>
-
 <head>
   <title>MapTiler JS SDK example</title>
   <style>
