+++ demos/03-sky-night.html
@@ -1,5 +1,5 @@
+<!DOCTYPE html>
 <html>
-
 <head>
   <title>MapTiler JS SDK example</title>
   <style>
