+++ demos/04-ready-event.html
@@ -1,3 +1,4 @@
+<!DOCTYPE html>
 <html>
 
 <head>
