+++ demos/05-transform-request.html
@@ -1,5 +1,5 @@
+<!DOCTYPE html>
 <html>
-
 <head>
   <title>MapTiler JS SDK example</title>
   <style>
