+++ demos/06-handling-webgl-errors.html
@@ -1,3 +1,4 @@
+<!DOCTYPE html>
 <html>
 
 <head>
