+++ demos/index.html
@@ -58,6 +58,7 @@
       <a href="04-ready-event.html?key=">Ready Event →</a>
       <a href="05-transform-request.html?key=">Transform request →</a>
       <a href="06-handling-webgl-errors.html?key=">Handling WebGL Errors →</a>
+      <a href="07-animated-routes.html?key=">Animated Routes →</a>
     </div>
   </div>
