+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@maptiler/sdk",
-  "version": "3.2.0",
+  "version": "3.3.0-rc1",
   "description": "The Javascript & TypeScript map SDK tailored for MapTiler Cloud",
   "author": "MapTiler",
   "module": "dist/maptiler-sdk.mjs",
@@ -78,7 +78,7 @@
   },
   "dependencies": {
     "@maplibre/maplibre-gl-style-spec": "^23.0.0",
-    "@maptiler/client": "^2.2.0",
+    "@maptiler/client": "^2.3.0",
     "events": "^3.3.0",
     "js-base64": "^3.7.7",
     "maplibre-gl": "^5.0.1",
