--- src/utils/MTAnimation/index.ts
@@ -1,4 +1,0 @@
-import MTAnimation from "./MTAnimation";
-
-export default MTAnimation;
-export * from "./types";
