+++ src/utils/MaptilerAnimation/MaptilerAnimation.ts
@@ -15,22 +15,22 @@ import EasingFunctions from "./easing";
  *                                   and must be updated manually by calling update(). Defaults to false if not specified.
  */
 export interface AnimationOptions {
-  // the keyframes to animate between
   keyframes: Keyframe[];
-  // the duration of the animation in milliseconds
   duration: number;
-  // the number of times to repeat the
-  // animation, 0 is no repeat, Infinity is infinite repeat
   iterations: number;
-  // if true, the animation will not be added to the
-  // animation manager and will need to be updated manually
-  // by calling update()
   manualMode?: boolean;
-
-  // the delay before the animation starts in milliseconds
   delay?: number;
 }
 
+/**
+ * A keyframe in the animation sequence where null or undefined values
+ * are interpolated to fill in the gaps between keyframes.
+ *
+ * @extends Keyframe
+ * @property {Record<string, number>} props - The properties to be animated at this keyframe.
+ * @property {EasingFunctionName} easing - The easing function to use for this keyframe.
+ * @property {string} id - A unique identifier for this keyframe.
+ */
 export type InterpolatedKeyFrame = Keyframe & {
   props: Record<string, number>;
   easing: EasingFunctionName;
@@ -40,12 +40,12 @@ export type InterpolatedKeyFrame = Keyframe & {
 /**
  * Animation controller for keyframe-based animation sequences.
  *
- * MTAnimation handles interpolation between keyframes, timing control,
+ * MaptilerAnimation handles interpolation between keyframes, timing control,
  * and event dispatching during animation playback.
  *
  * @example
  * ```typescript
- * const animation = new MTAnimation({
+ * const animation = new MaptilerAnimation({
  *   keyframes: [
  *     { delta: 0, props: { x: 0, y: 0 } },
  *     { delta: 0.5, props: { x: 50, y: 20 } },
@@ -72,53 +72,68 @@ export type InterpolatedKeyFrame = Keyframe & {
  *
  * When not using manualMode, animations are automatically added to the AnimationManager.
  */
-export default class MTAnimation {
+export default class MaptilerAnimation {
   private playing: boolean = false;
 
+  /**
+   * Indicates if the animation is currently playing
+   * @returns {boolean} - true if the animation is playing, false otherwise
+   */
   get isPlaying() {
     return this.playing;
   }
 
-  // the number of times to repeat the animation
-  // 0 is no repeat, Infinity is infinite repeat
+  /**
+   * The number of times to repeat the animation
+   * 0 is no repeat, Infinity is infinite repeat
+   */
   private iterations: number;
 
   private currentIteration: number = 0;
 
-  // an array of keyframes animations to interpolate between
+  /** An array of keyframes animations to interpolate between */
   private keyframes: InterpolatedKeyFrame[];
 
+  /** The current keyframe id */
   private currentKeyframe?: string;
 
-  // the duration of the animation in milliseconds
+  /**The duration of the animation in milliseconds (when playbackRate === 1) */
   readonly duration: number;
 
-  // the duration of the animation affected by the playback rate
-  // if playback rate is 2, the effective duration is double
+  /**
+   * The duration of the animation affected by the playback rate
+   * if playback rate is 2, the effective duration is double
+   */
   private effectiveDuration: number;
 
-  // the rate at which the animation is playing
+  /** the rate at which the animation is playing */
   private playbackRate: number;
 
-  // the current time in milliseconds
+  /** the current time in milliseconds */
   private currentTime: number;
 
-  // 0 start of the animation, 1 end of the animation
+  /** 0 start of the animation, 1 end of the animation */
   private currentDelta: number;
 
+  /** The time at which the animation started */
   private animationStartTime: number = 0;
 
+  /** The time at which the last frame was rendered */
   private lastFrameAt: number = 0;
 
+  /** The delay before the animation starts */
   private delay: number = 0;
 
+  /** The timeout ID for the delay before the animation starts */
   private delayTimeoutID?: number;
 
+  /** The listeners added for each event */
   private listeners: AnimationEventListenersRecord = Object.values(AnimationEventTypes).reduce((acc, type) => {
     acc[type] = [];
     return acc;
   }, {} as AnimationEventListenersRecord);
 
+  /** The props from the previous frame */
   private previousProps!: Record<string, number>;
 
   constructor({ keyframes, duration, iterations, manualMode, delay }: AnimationOptions) {
@@ -335,7 +350,7 @@ export default class MTAnimation {
         const t = (this.currentDelta - current.delta) / (next.delta - current.delta);
 
         // get the easing function to use
-        const easingFunc = EasingFunctions[current.easing];
+        const easingFunc = EasingFunctions[current.easing] ?? ((n: number) => n);
 
         // get the alpha value from the easing function
         // this value is the amount to interpolate between
@@ -538,7 +553,7 @@ export default class MTAnimation {
    * @returns A new animation instance with the same properties as this one
    */
   clone() {
-    return new MTAnimation({
+    return new MaptilerAnimation({
       keyframes: this.keyframes,
       duration: this.duration,
       iterations: this.iterations,
