+++ src/utils/MaptilerAnimation/easing.ts
