+++ src/utils/MaptilerAnimation/index.ts
@@ -1,0 +1,6 @@
+import MaptilerAnimation from "./MaptilerAnimation";
+
+export default MaptilerAnimation;
+export * from "./types";
+export * from "./easing";
+export * from "./animation-helpers";
